+++ src/main/java/JOO/jooshop/product/model/ProductRankResponseDto.java
@@ -7,6 +7,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -17,7 +18,7 @@ public class ProductRankResponseDto {
     private Long productId;           // 상품 고유 번호 (Primary Key)
     private ProductType productType;  // 상품의 유형 (남성/여성/유니섹스 등 ENUM)
     private String productName;       // 상품 이름
-    private Integer price;            // 상품 가격
+    private BigDecimal price;            // 상품 가격
     private Long wishListCount;       // 위시리스트에 추가된 횟수 (인기 상품 기준에 사용될 수 있음)
     private Boolean isDiscount;       // 할인 여부 (true = 할인중, false = 비할인)
     private Integer discountRate;     // 할인율 (퍼센트로 저장, 예: 20 -> 20%)
